Remove rusqlite in favor of libsqlite3
