+++ crates/but-db/Cargo.toml
@@ -10,18 +10,18 @@ doctest = false
 test = false
 
 [dependencies]
-diesel = { version = "2.2.11", features = [
+diesel = { version = "2.2.10", features = [
     "sqlite",
     "returning_clauses_for_sqlite_3_35",
     "chrono",
 ] }
+# Bundled libsqlite3 for Windows
+libsqlite3-sys = { version = "0.29", features = ["bundled"] }
+bitflags = "2.9.1"
 serde = { workspace = true, features = ["std"] }
 anyhow = "1.0.98"
 diesel_migrations = { version = "2.0.0", features = ["sqlite"] }
 chrono = { version = "0.4.41", features = ["serde"] }
-# Needed for bundled libsqlite3 for Windows and hooks
-rusqlite = { version = "0.35.0", default-features = false, features = ["hooks", "bundled"]}
-bitflags = "2.9.1"
 
 [dev-dependencies]
 tempfile = "3.20.0"
